Move context test fixture to avoid import cycle
